cambio atributo bd precio plan, home conexion bd en cards

+++ webpage/models.py
@@ -2,7 +2,7 @@ from django.db import models
 
 class Plan(models.Model):
     nombre = models.CharField(max_length=100, unique=True)
-    precio = models.DecimalField(max_digits=10, decimal_places=2)
+    precio = models.PositiveIntegerField()
 
     def __str__(self):
         return self.nombre
